+++ changelog.md
@@ -5,6 +5,8 @@
 ### 1.6.2, 2020-03-10
 
 * fixed: activating the must-use plugin throws an error if the mu-plugins folder is missing
+* changed: can now enable both a Toolbar indicator and a site-wide admin notice when emails are disabled
+* changed: filter hook `disable_emails_indicator` also accepts 'notice_toolbar' to enable both notice and Toolbar indicator
 
 ### 1.6.1, 2019-12-21